extract print function to separate file
